add start/1 to webmachine_ws, store web server in env
Add webmachine_ws:start/1 to provide a portable start function for
applications that don't want to be tied to a specific web server startup
function.

Make webmachine_ws:get_webserver_mod/0 first check the application env for
the name of the underlying web server. If not found, check the operating
system process environment, and if found there, store it, or a default of
mochiweb if not found, into the application env.